Implement a way to remove sectors
